// src/api/interfaces/ISeries.ts
export interface ISeries{
  id: string;
  title: string;
  poster: string; 
  year: string;
  seasons: number;
  genres: string[];      
  rating: number;    
  synopsis: string;     
  authors: string[];     
  director: string;
  writers: string[];
  country: string;
  releaseDate: string;
  promo: string;
};